refactor(condo): DOMA-5050 extract card content to component
