+++ apps/condo/lang/en/en.json
@@ -503,9 +503,12 @@
   "pages.condo.property.warning.modal.Title": "Do not save property?",
   "pages.condo.property.id.DownloadExcelLabel": "Download file",
   "pages.condo.property.ImportTitle": "properties",
-  "pages.condo.property.id.PropertyInformationTitle": "About property",
-  "pages.condo.property.id.PropertyReportTitle": "Property report",
-  "pages.condo.property.id.PropertyReportDescription": "If You are a client of Sber, You're able to present a cool report of property expenses to Your residents. Otherwise, become one by sending a request",
+  "pages.condo.property.id.propertyInformationTitle": "About property",
+  "pages.condo.property.id.propertyReportTitle": "Property report",
+  "pages.condo.property.id.propertyReportDescription": "If You are a client of Sber, You're able to present a cool report of property expenses to Your residents. Otherwise, become one by sending a request",
+  "pages.condo.property.id.propertyReportComingSoonTitle": "«Property report» coming soon",
+  "pages.condo.property.id.propertyReportComingSoonSubTitle": "This is a new feature for Sber customers.",
+  "pages.condo.property.id.propertyReportComingSoonDescription": "With this you can send a clear and detailed report with important figures about the expenses and income of the house to your residents. Details coming soon.",
   "pages.condo.property.id.becomeSberClientTitle": "Become Sber client",
   "pages.condo.property.id.setupReportTitle": "Setup report",
   "pages.condo.meter.create.AddMeterReadingPlaceholder": "Enter value",
