+++ apps/condo/lang/ru/ru.json
@@ -503,9 +503,12 @@
   "pages.condo.property.warning.modal.Title": "Не сохранять дом?",
   "pages.condo.property.id.DownloadExcelLabel": "Скачать файл",
   "pages.condo.property.ImportTitle": "домов",
-  "pages.condo.property.id.PropertyInformationTitle": "О доме",
-  "pages.condo.property.id.PropertyReportTitle": "Отчет по дому",
-  "pages.condo.property.id.PropertyReportDescription": "Если вы клиент Сбера, жителям ваших домов доступен классный отчет. Если не клиент — станьте им, оставив заявку",
+  "pages.condo.property.id.propertyInformationTitle": "О доме",
+  "pages.condo.property.id.propertyReportTitle": "Отчет по дому",
+  "pages.condo.property.id.propertyReportDescription": "Если вы клиент Сбера, жителям ваших домов доступен классный отчет. Если не клиент — станьте им, оставив заявку",
+  "pages.condo.property.id.propertyReportComingSoonTitle": "Скоро здесь появится «Отчет по дому»",
+  "pages.condo.property.id.propertyReportComingSoonSubTitle": "Это новая функция для клиентов Сбера.",
+  "pages.condo.property.id.propertyReportComingSoonDescription": "С ее помощью жителям можно отправить понятный и подробный отчет с важными цифрами о расходах и доходах дома. Подробности — скоро.",
   "pages.condo.property.id.becomeSberClientTitle": "Стать клиентом Сбера",
   "pages.condo.property.id.setupReportTitle": "Настроить отчет",
   "pages.condo.meter.create.AddMeterReadingPlaceholder": "Введите значение",
